refactor(cli): put every subcommand under commands/
Move cmd/* and gc.rs into commands/, extract the inlined rule/skill/plan/
execpolicy/version handlers, and keep commands.rs as clap types plus thin
dispatch so new commands have one place to land.

Fixes #1954
