--- crates/harness-cli/src/cmd/mod.rs
@@ -1,2 +1,0 @@
-pub mod mcp_server;
-pub mod pr;
