+++ crates/harness-cli/src/commands.rs
@@ -10,12 +10,20 @@ use tracing_subscriber::fmt::writer::MakeWriter;
 
 mod eval;
 mod exec;
+mod execpolicy;
+mod gc;
+mod mcp_server;
+mod plan;
+mod pr;
 mod reconcile;
+mod rule;
 mod runtime;
 #[cfg(test)]
 mod runtime_log_tests;
 mod serve;
+mod skill;
 mod status;
+mod version;
 
 const RUNTIME_LOG_PREFIX: &str = "harness-serve-";
 const RUNTIME_LOG_SUFFIX: &str = ".log";
@@ -341,7 +349,7 @@ pub enum RuntimeBreakerCommand {
     },
 }
 
-fn configured_rule_engine(
+pub(crate) fn configured_rule_engine(
     config: &harness_core::config::HarnessConfig,
 ) -> harness_rules::engine::RuleEngine {
     let mut engine = harness_rules::engine::RuleEngine::new();
@@ -351,18 +359,6 @@ fn configured_rule_engine(
         config.rules.requirements_path.clone(),
     );
     engine
-}
-
-fn configured_skill_store(
-    config: &harness_core::config::HarnessConfig,
-) -> anyhow::Result<harness_skills::store::SkillStore> {
-    let project_root = std::env::current_dir()?;
-    let mut store = harness_skills::store::SkillStore::new()
-        .with_persist_dir(config.server.data_dir.join("skills"))
-        .with_discovery(&project_root);
-    store.load_builtin();
-    store.discover()?;
-    Ok(store)
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -825,7 +821,7 @@ pub async fn run(cli: Cli) -> anyhow::Result<()> {
         }
 
         Command::McpServer => {
-            crate::cmd::mcp_server::run(config.clone()).await?;
+            mcp_server::run(config.clone()).await?;
         }
 
         Command::Exec {
@@ -859,145 +855,30 @@ pub async fn run(cli: Cli) -> anyhow::Result<()> {
         }
 
         Command::Gc { cmd } => {
-            crate::gc::run_gc(cmd, &config).await?;
+            gc::run_gc(cmd, &config).await?;
         }
 
         Command::Rule { cmd } => {
-            match cmd {
-                RuleCommand::Load { project } => {
-                    let mut engine = configured_rule_engine(&config);
-                    engine.load(&project)?;
-                    println!("Loaded {} rules", engine.rules().len());
-                }
-                RuleCommand::Check { project, auto_fix } => {
-                    let mut engine = configured_rule_engine(&config);
-                    engine.load(&project)?;
-                    let violations = engine.scan(&project).await?;
-                    // Persist rule scan results for observability/GC even when running via CLI.
-                    match harness_observe::event_store::EventStore::with_policies_and_otel_with_database_url(
-                        &config.server.data_dir,
-                        config.server.database_url.as_deref(),
-                        config.observe.session_renewal_secs,
-                        config.observe.log_retention_days,
-                        &config.otel,
-                    )
-                    .await
-                    {
-                        Ok(store) => {
-                            store.persist_rule_scan(&project, &violations).await;
-                            store.shutdown().await;
-                        }
-                        Err(e) => tracing::warn!(
-                            "Failed to initialize event store, rule scan not persisted: {e}"
-                        ),
-                    }
-                    if violations.is_empty() {
-                        println!("No violations found");
-                    } else {
-                        for v in &violations {
-                            println!(
-                                "{:?} {}:{} [{}] {}",
-                                v.severity,
-                                v.file.display(),
-                                v.line.unwrap_or(0),
-                                v.rule_id,
-                                v.message
-                            );
-                        }
-                        if auto_fix {
-                            let fixed = engine.apply_fixes(&violations, &project)?;
-                            println!("Auto-fixed {fixed} file(s)");
-                        }
-                    }
-                }
-            }
+            rule::run(cmd, &config).await?;
         }
 
-        Command::ExecPolicy { cmd } => match cmd {
-            ExecPolicyCommand::Check {
-                rules,
-                requirements,
-                resolve_host_executables,
-                pretty,
-                command,
-            } => {
-                let mut engine = configured_rule_engine(&config);
-                let policy_paths = if rules.is_empty() {
-                    config.rules.exec_policy_paths.clone()
-                } else {
-                    rules
-                };
-                if policy_paths.is_empty() {
-                    anyhow::bail!(
-                        "no execpolicy rules supplied; pass --rules or set rules.exec_policy_paths"
-                    );
-                }
+        Command::ExecPolicy { cmd } => {
+            execpolicy::run(cmd, &config)?;
+        }
 
-                engine.load_exec_policy_files(&policy_paths)?;
-                if let Some(path) = requirements {
-                    engine.load_requirements_toml(&path)?;
-                } else {
-                    engine.load_configured_requirements()?;
-                }
-
-                let result = engine.check_command_policy(
-                    &command,
-                    &harness_rules::exec_policy::MatchOptions {
-                        resolve_host_executables,
-                    },
-                );
-                let rendered = if pretty {
-                    serde_json::to_string_pretty(&result)?
-                } else {
-                    serde_json::to_string(&result)?
-                };
-                println!("{rendered}");
-            }
-        },
-
-        Command::Skill { cmd } => match cmd {
-            SkillCommand::List { query } => {
-                let store = configured_skill_store(&config)?;
-                let skills = if let Some(q) = query {
-                    store.search(&q).into_iter().cloned().collect::<Vec<_>>()
-                } else {
-                    store.list().to_vec()
-                };
-                for s in &skills {
-                    println!("{} [{}]: {}", s.name, s.id, s.description);
-                }
-                if skills.is_empty() {
-                    println!("No skills found");
-                }
-            }
-            SkillCommand::Create { name, file } => {
-                let content = std::fs::read_to_string(&file)?;
-                let mut store = configured_skill_store(&config)?;
-                store.create(name.clone(), content);
-                println!("Created skill: {name}");
-            }
-            SkillCommand::Delete { skill_id } => {
-                let mut store = configured_skill_store(&config)?;
-                let deleted = if let Some(skill) = store.get_by_name(&skill_id).cloned() {
-                    store.delete(&skill.id)
-                } else {
-                    store.delete(&harness_core::types::SkillId::from_str(&skill_id))
-                };
-                println!("Deleted skill: {deleted}");
-            }
-        },
+        Command::Skill { cmd } => {
+            skill::run(cmd, &config)?;
+        }
 
         Command::Pr { cmd } => match cmd {
             PrCommand::Fix { issue, args } => {
-                crate::cmd::pr::fix(&config, issue, args.wait, args.max_rounds, args.project)
-                    .await?;
+                pr::fix(&config, issue, args.wait, args.max_rounds, args.project).await?;
             }
             PrCommand::Loop { pr, args } => {
-                crate::cmd::pr::loop_pr(&config, pr, args.wait, args.max_rounds, args.project)
-                    .await?;
+                pr::loop_pr(&config, pr, args.wait, args.max_rounds, args.project).await?;
             }
             PrCommand::Review { pr, args } => {
-                crate::cmd::pr::review(&config, pr, args.provider, args.base, args.project).await?;
+                pr::review(&config, pr, args.provider, args.base, args.project).await?;
             }
         },
 
@@ -1005,54 +886,12 @@ pub async fn run(cli: Cli) -> anyhow::Result<()> {
             eval::run(cmd).await?;
         }
 
-        Command::Plan { cmd } => match cmd {
-            PlanCommand::Init { spec } => {
-                let content = std::fs::read_to_string(&spec)?;
-                let project_root = std::env::current_dir()?;
-                let plan = harness_exec::plan::ExecPlan::from_spec(&content, &project_root)?;
-                let md = plan.to_markdown();
-                let out_path = format!("exec-plan-{}.md", plan.id);
-                std::fs::write(&out_path, &md)?;
-                println!("Created ExecPlan: {out_path}");
-            }
-            PlanCommand::Status { plan } => {
-                if std::path::Path::new(&plan).exists() {
-                    let content = std::fs::read_to_string(&plan)?;
-                    let p = harness_exec::plan::ExecPlan::from_markdown(&content)?;
-                    println!("Plan: {}", p.purpose);
-                    println!("Status: {:?}", p.status);
-                    let done = p.progress.iter().filter(|m| m.completed).count();
-                    println!("Progress: {}/{}", done, p.progress.len());
-                } else {
-                    println!("Plan file not found: {plan}");
-                }
-            }
-        },
+        Command::Plan { cmd } => {
+            plan::run(cmd)?;
+        }
 
         Command::Version => {
-            let cargo_toml_path = std::env::current_dir()?.join("Cargo.toml");
-            if !cargo_toml_path.exists() {
-                anyhow::bail!("Cargo.toml not found in current directory");
-            }
-            let content = std::fs::read_to_string(&cargo_toml_path)?;
-            let parsed: toml::Value = toml::from_str(&content)?;
-
-            if let Some(version) = parsed
-                .get("workspace")
-                .and_then(|w| w.get("package"))
-                .and_then(|p| p.get("version"))
-                .and_then(|v| v.as_str())
-            {
-                println!("Current version: {}", version);
-            } else if let Some(version) = parsed
-                .get("package")
-                .and_then(|p| p.get("version"))
-                .and_then(|v| v.as_str())
-            {
-                println!("Current version: {}", version);
-            } else {
-                anyhow::bail!("Version field not found in Cargo.toml");
-            }
+            version::run()?;
         }
 
         Command::Status {
