+++ crates/harness-cli/src/main.rs
@@ -3,9 +3,7 @@
 
 use clap::Parser;
 
-mod cmd;
 mod commands;
-mod gc;
 
 fn main() -> anyhow::Result<()> {
     let cli = commands::Cli::parse();
